refactor: split main app into modules
